+++ packages/ibm-products/src/components/CreateFullPage/CreateFullPage.test.js
@@ -8,7 +8,7 @@
 import React from 'react';
 import { render, screen, waitFor, act } from '@testing-library/react'; // https://testing-library.com/docs/react-testing-library/intro
 import userEvent from '@testing-library/user-event';
-import { pkg } from '../../settings';
+import { carbon, pkg } from '../../settings';
 import uuidv4 from '../../global/js/utils/uuidv4';
 import {
   expectWarn,
@@ -120,6 +120,7 @@ const renderCreateFullPage = ({
         title="Title 2"
         description="2"
         fieldsetLegendText="2"
+        invalid={false}
       >
         {stepFormField}
       </CreateFullPageStep>
@@ -127,6 +128,7 @@ const renderCreateFullPage = ({
         title="Title 3"
         description="3"
         onNext={rejectOnSubmitNext ? finalStepOnNextRejectFn : finalOnNextFn}
+        invalid
       >
         {stepFormField}
       </CreateFullPageStep>
@@ -171,6 +173,18 @@ const renderFullPageWithStepChildrenOutside = ({ ...rest } = {}) =>
   );
 
 describe(componentName, () => {
+  const { ResizeObserver } = window;
+  beforeEach(() => {
+    window.ResizeObserver = jest.fn().mockImplementation(() => ({
+      observe: jest.fn(),
+      unobserve: jest.fn(),
+      disconnect: jest.fn(),
+    }));
+  });
+  afterEach(() => {
+    window.ResizeObserver = ResizeObserver;
+  });
+
   // Currently fails due to https://github.com/carbon-design-system/carbon/issues/14135 regarding focusable button
   it.skip('has no accessibility violations', async () => {
     const { container } = renderComponent({ ...defaultFullPageProps });
@@ -513,5 +527,59 @@ describe(componentName, () => {
     expect(
       createFullPageSteps[0].classList.contains(`.${blockClass}__step-fieldset`)
     );
+  });
+
+  it('renders a header if title is provided ', () => {
+    const { container } = renderCreateFullPage({
+      ...defaultFullPageProps,
+      title: 'Page title',
+    });
+    const headerSelector = container.querySelector(`.${blockClass}__header`);
+
+    expect(headerSelector).toBeInTheDocument();
+  });
+
+  it('renders a header if breadcrumbs are provided', () => {
+    const { container } = renderCreateFullPage({
+      ...defaultFullPageProps,
+      breadcrumbs: [
+        { key: '0', href: '/', label: 'Home page' },
+        { key: '1', href: '/', label: 'Application name' },
+      ],
+      breadcrumbsOverflowAriaLabel: 'breadcrumbs overflow aria-label',
+    });
+    const headerSelector = container.querySelector(`.${blockClass}__header`);
+
+    expect(headerSelector).toBeInTheDocument();
+  });
+
+  it("doesn't render a header if title or breadcrumbs are not provided  ", () => {
+    const { container } = renderCreateFullPage({
+      ...defaultFullPageProps,
+    });
+    const headerSelector = container.querySelector(`.${blockClass}__header`);
+    expect(headerSelector).not.toBeInTheDocument();
+  });
+
+  it('renders an error icon if the step invalid prop is set to true', async () => {
+    renderCreateFullPage({
+      ...defaultFullPageProps,
+    });
+
+    expect(
+      screen
+        .getByRole('button', { description: 'Title 1' })
+        .querySelector(`.${carbon.prefix}--progress__warning`)
+    ).not.toBeInTheDocument();
+    expect(
+      screen
+        .getByRole('button', { description: 'Title 2' })
+        .querySelector(`.${carbon.prefix}--progress__warning`)
+    ).not.toBeInTheDocument();
+    expect(
+      screen
+        .getByRole('button', { description: 'Title 3' })
+        .querySelector(`.${carbon.prefix}--progress__warning`)
+    ).toBeInTheDocument();
   });
 });