+++ packages/ibm-products/src/components/Datagrid/Datagrid/addons/CustomizeColumns/DraggableItemsList.js
@@ -11,7 +11,6 @@ import { Checkbox } from '@carbon/react';
 import { isColumnVisible } from './common';
 import DraggableElement from '../../DraggableElement';
 import { pkg } from '../../../../../settings';
-import getColTitle from '../../../utils/getColTitle';
 
 import {
   DndContext,
@@ -25,6 +24,7 @@ import {
   SortableContext,
   verticalListSortingStrategy,
 } from '@dnd-kit/sortable';
+import { getNodeTextContent } from '../../../../../global/js/utils/getNodeTextContent';
 
 const blockClass = `${pkg.prefix}--datagrid`;
 const matchedColsById = (col1, col2) => col1 && col2 && col1.id === col2.id;
@@ -40,22 +40,38 @@ export const DraggableItemsList = ({
   const draggableClass = `${blockClass}__draggable-item`;
   const visibleCols = columns
     // hide the columns without Header, e.g the sticky actions, spacer
-    .filter((colDef) => {
-      return !!getColTitle(colDef);
-    })
+    .filter((colDef) => getNodeTextContent(colDef.Header).trim().length !== 0)
     .filter(Boolean)
     .filter((colDef) => !colDef.isAction)
+    .filter((colDef) => colDef.id !== 'spacer')
     .filter((colDef) => {
       return (
         filterString.length === 0 ||
-        (getColTitle(colDef)?.toLowerCase().includes(filterString) &&
-          colDef.id !== 'spacer')
+        getNodeTextContent(colDef.Header).toLowerCase().includes(filterString)
       );
     });
+
+  const getUpdatedDragCols = () => {
+    const tempCols = [...visibleCols];
+    tempCols.forEach((col) => {
+      if (col.sticky) {
+        col.disabled = true;
+      }
+    });
+    return tempCols;
+  };
+  const updatedDragCols = getUpdatedDragCols();
 
   // let localRefCopy;
   const handleDragEnd = (event) => {
     const { active, over } = event;
+
+    // Stop any re-ordering updates if the destination column is disabled
+    // ie: it is a frozen column.
+    const destOverCol = updatedDragCols.filter((item) => item.id === over.id);
+    if (destOverCol?.length && destOverCol[0]?.disabled) {
+      return;
+    }
 
     const fromVisibleIndex = columns.findIndex((col) =>
       matchedColsById(col, active)
@@ -63,11 +79,14 @@ export const DraggableItemsList = ({
     const toVisibleIndex = columns.findIndex((col) =>
       matchedColsById(col, over)
     );
-    const colTitle = getColTitle(visibleCols[fromVisibleIndex]);
+
+    const colTitle = getNodeTextContent(
+      updatedDragCols[fromVisibleIndex].Header
+    );
 
     setAriaRegionText(
       `${colTitle} dropped. New position ${toVisibleIndex + 1} of ${
-        visibleCols.length
+        updatedDragCols.length
       }.`
     );
 
@@ -80,14 +99,14 @@ export const DraggableItemsList = ({
   const handleDragStart = (event) => {
     const { active } = event;
 
-    const fromIndex = visibleCols.findIndex((col) =>
+    const fromIndex = updatedDragCols.findIndex((col) =>
       matchedColsById(col, active)
     );
-    const colTitle = getColTitle(visibleCols[fromIndex]);
+    const colTitle = getNodeTextContent(updatedDragCols[fromIndex].Header);
 
     setAriaRegionText(
       `${colTitle} grabbed. Current position ${fromIndex + 1} of ${
-        visibleCols.length
+        updatedDragCols.length
       }.`
     );
   };
@@ -95,17 +114,19 @@ export const DraggableItemsList = ({
   const handleDragUpdate = (event) => {
     const { active, over } = event;
 
-    const fromIndex = visibleCols.findIndex((col) =>
+    const fromIndex = updatedDragCols.findIndex((col) =>
       matchedColsById(col, active)
     );
-    const toIndex = visibleCols.findIndex((col) => matchedColsById(col, over));
+    const toIndex = updatedDragCols.findIndex((col) =>
+      matchedColsById(col, over)
+    );
 
-    const colTitle = getColTitle(visibleCols[fromIndex]);
+    const colTitle = getNodeTextContent(updatedDragCols[fromIndex].Header);
 
     setAriaRegionText(
       `${colTitle} grabbed. Original position ${fromIndex + 1}, new position ${
         toIndex + 1
-      } of ${visibleCols.length}.`
+      } of ${updatedDragCols.length}.`
     );
   };
 
@@ -171,15 +192,15 @@ export const DraggableItemsList = ({
             <div
               className={`${blockClass}__draggable-underlay-item`}
               key={colDef.id}
-            ></div>
+            />
           ))}
         </div>
         <SortableContext
-          items={visibleCols}
+          items={updatedDragCols}
           strategy={verticalListSortingStrategy}
         >
           {visibleCols.map((colDef) => {
-            const colHeaderTitle = getColTitle(colDef);
+            const colHeaderTitle = getNodeTextContent(colDef.Header);
             const searchString = new RegExp('(' + filterString + ')');
             const res = filterString.length
               ? colHeaderTitle.toLowerCase().split(searchString)
@@ -214,7 +235,7 @@ export const DraggableItemsList = ({
                   <div
                     dangerouslySetInnerHTML={{ __html: highlightedText }}
                     className={`${blockClass}__customize-columns-checkbox-visible-label`}
-                  ></div>
+                  />
                 }
               </>
             );