+++ packages/ibm-products/src/components/Datagrid/useSelectRows.js
@@ -76,7 +76,7 @@ const SelectRow = (datagridState) => {
     return () => window.removeEventListener('resize', updateSize);
   }, []);
 
-  const selectDisabled = isFetching || row.getRowProps().selectDisabled;
+  const selectDisabled = isFetching || row.getRowProps().disabled;
   const { onChange, ...selectProps } = row.getToggleRowSelectedProps();
   const cellProps = cell.getCellProps();
   const isFirstColumnStickyLeft =