+++ packages/ibm-products/src/components/EditTearsheetNarrow/EditTearsheetNarrow.stories.js
@@ -291,6 +291,5 @@ export const withMultipleForms = prepareStory(WithValidationTemplate, {
   storyName: withMultipleFormsStoryName,
   args: {
     ...defaultStoryProps,
-    multiForm: true,
   },
 });