+++ packages/ibm-products/src/components/FilterSummary/FilterSummary.js
@@ -21,6 +21,7 @@ let FilterSummary = React.forwardRef(
       clearFilters = () => {},
       filters = [],
       renderLabel = null,
+      overflowType = 'default',
     },
     ref
   ) => {
@@ -40,6 +41,7 @@ let FilterSummary = React.forwardRef(
           allTagsModalTitle="All tags"
           showAllTagsLabel="View all tags"
           tags={tagFilters}
+          overflowType={overflowType}
         />
         <Button kind="ghost" size="sm" onClick={clearFilters}>
           {clearFiltersText}
@@ -57,6 +59,7 @@ FilterSummary.propTypes = {
   clearFilters: PropTypes.func.isRequired,
   clearFiltersText: PropTypes.string,
   filters: PropTypes.arrayOf(PropTypes.object).isRequired,
+  overflowType: PropTypes.oneOf(['default', 'tag']),
   renderLabel: PropTypes.func,
 };