+++ packages/ibm-products/src/components/Tearsheet/Tearsheet.js
@@ -23,6 +23,8 @@ import { ActionSet } from '../ActionSet';
 
 import { tearsheetHasCloseIcon, TearsheetShell } from './TearsheetShell';
 
+import { portalType } from './TearsheetShell';
+
 const componentName = 'Tearsheet';
 
 // NOTE: the component SCSS is not imported here: it is rolled up separately.
@@ -204,9 +206,9 @@ Tearsheet.propTypes = {
   open: PropTypes.bool,
 
   /**
-   * The DOM node the tearsheet should be rendered within. Defaults to document.body.
+   * The DOM element that the tearsheet should be rendered within. Defaults to document.body.
    */
-  portalTarget: PropTypes.node,
+  portalTarget: portalType,
 
   /**
    * Specify a CSS selector that matches the DOM element that should be focused when the Modal opens